+++ sscCdi/caterete/cat_ptycho_processing.py
@@ -42,6 +42,11 @@ def cat_ptychography(input_dict,restoration_dict,restored_data_info, filepaths, 
     frame_index = input_dict["projections"]
     corrected_positions_list = []
 
+    # defining flag for multiple initial probe and/or objects from previous ptycho
+    run_again     = input_dict["mult_obj_probe"]["run_again"]
+    initial_obj   = input_dict["mult_obj_probe"]["mult_obj"]
+    initial_probe = input_dict["mult_obj_probe"]["mult_probe"]
+
     if strategy == "serial":
 
         for file_number_index, filename in enumerate(filenames):
@@ -85,10 +90,15 @@ def cat_ptychography(input_dict,restoration_dict,restored_data_info, filepaths, 
             probe_positions, angle = read_probe_positions(input_dict, acquisitions_folder,filename , DPs.shape)
             print(f"\tFinished reading probe positions. Shape: {probe_positions.shape}")
 
-            if file_number == 0: # compute object size, object pixel size for the first frame and use it in all 3D ptycho
+            if file_number_index == 0:
                 input_dict = set_object_shape(input_dict, DPs.shape, probe_positions)
                 sinogram              = np.zeros((total_number_of_angles,input_dict["object_shape"][0],input_dict["object_shape"][1]),dtype=np.complex64) 
-                probes                = np.zeros((total_number_of_angles,input_dict["incoherent_modes"],DPs.shape[-2],DPs.shape[-1]),dtype=np.complex64)
+                
+                if input_dict["incoherent_modes"] < 1:
+                    incoherent_modes = 1
+                else:
+                    incoherent_modes = input_dict["incoherent_modes"]
+                probes = np.zeros((total_number_of_angles,incoherent_modes,DPs.shape[-2],DPs.shape[-1]),dtype=np.complex64)
 
                 print(f"\tInitial object shape: {sinogram.shape}\t Initial probe shape: {probes.shape}")
 
@@ -113,6 +123,21 @@ def cat_ptychography(input_dict,restoration_dict,restored_data_info, filepaths, 
                 
                 angle = np.array([file_number_index,0,angle,angle*180/np.pi])
 
+            if run_again:
+                print("Second ptycho run")
+                if initial_obj and initial_probe:
+                    print("Running with multiple initial objects and probes")
+                    sinogram[file_number_index, :, :], probes[file_number_index, :, :], error, corrected_positions = call_GB_ptychography(input_dict,DPs,probe_positions,initial_obj=sinogram[file_number_index, :, :], initial_probe=probes[file_number_index, :, :]) # run ptycho
+                elif initial_probe:
+                    print("Running with multiple initial probes")
+                    sinogram[file_number_index, :, :], probes[file_number_index, :, :], error, corrected_positions = call_GB_ptychography(input_dict,DPs,probe_positions,initial_probe=probes[file_number_index, :, :]) # run ptycho
+                elif initial_obj:
+                    print("Running with multiple initial objects")
+                    sinogram[file_number_index, :, :], probes[file_number_index, :, :], error, corrected_positions = call_GB_ptychography(input_dict,DPs,probe_positions,initial_obj=sinogram[file_number_index, :, :]) # run ptycho
+                
+                if corrected_positions is not None:
+                    corrected_positions_list[file_number_index] = corrected_positions[:,0,0:2]
+
             """ Save single frame of object and probe to temporary folder"""
 
             np.save(os.path.join(input_dict["temporary_output_recons"],f"{file_number:04d}_object.npy"),sinogram[file_number_index])
@@ -133,7 +158,7 @@ def cat_ptychography(input_dict,restoration_dict,restored_data_info, filepaths, 
             pi540D.ioCleanM_Backward540D( restoration_dict, restored_data_info )
 
 
-    return input_dict, sinogram, probes, probe_positions, corrected_positions_list
+    return input_dict, sinogram, probes, probe_positions
 
 
 ##### ##### ##### #####                  DATA PREPARATION                 ##### ##### ##### ##### #####